+++ src/FSMDetector.cpp
@@ -12,10 +12,10 @@
 namespace targetfinder {
     class FSMDetector : public MarkerDetector {
         public:
-            static constexpr double MARKER_ASPECT_TOLERANCE = 0.9;
+            static constexpr float MARKER_ASPECT_TOLERANCE = 0.9;
             static constexpr int ROW_STEP = 1;
-            static constexpr double MIN_THRESHOLD = 25.0;
-            static constexpr double ALPHA = 0.00001;
+            static constexpr float MIN_THRESHOLD = 25.0;
+            static constexpr float ALPHA = 0.00001;
 
             void setMinThreshold(int thresh) {
                 this->min_threshold = thresh;
@@ -29,15 +29,15 @@ namespace targetfinder {
                 this->min_length = length;
             }
 
-            void setSmoothness(double smoothness) {
+            void setSmoothness(float smoothness) {
                 this->alpha = smoothness;
             }
 
-            void setTolerance(double tolerance) {
+            void setTolerance(float tolerance) {
                 this->tolerance = tolerance;
             }
 
-            void setMarkerAspectTolerance(double tolerance) {
+            void setMarkerAspectTolerance(float tolerance) {
                 this->marker_aspect_tolerance = tolerance;
             }
 
@@ -59,12 +59,12 @@ namespace targetfinder {
                         this->min_length
                 );       // Vertically-scanning state machine
 
-                double aspect = _aspect(input.cols, input.rows);
-                double alpha1 = ((1.0 / ((double) input.cols / this->alpha)));
-                double global_delta = 0.0;
+                float aspect = _aspect(input.cols, input.rows);
+                float alpha1 = ((1.0 / ((float) input.cols / this->alpha)));
+                float global_delta = 0.0;
                 int last_center_x = 0;
                 Marker *m = nullptr;
-                double threshold = 0.0;
+                float threshold = 0.0;
                 unsigned char horiz_last = 0;
                 unsigned char horiz = 0;
                 int horiz_diff = 0;
@@ -190,9 +190,9 @@ namespace targetfinder {
                                 }
                                 if (m2) {
                                     int min_pixels = (int) ((this->min_length * StateMachine::NUM_STATES) * this->tolerance);
-                                    double marker_aspect = _aspect(m->xlength(), m->ylength());
-                                    double min_marker_aspect = 1.0 - this->marker_aspect_tolerance;
-                                    double max_marker_aspect = 1.0 + this->marker_aspect_tolerance;
+                                    float marker_aspect = _aspect(m->xlength(), m->ylength());
+                                    float min_marker_aspect = 1.0 - this->marker_aspect_tolerance;
+                                    float max_marker_aspect = 1.0 + this->marker_aspect_tolerance;
                                     if (m->ylength() >= min_pixels
                                         && m->xlength() >= min_pixels
                                         && marker_aspect >= min_marker_aspect
@@ -216,9 +216,9 @@ namespace targetfinder {
             int row_step = ROW_STEP;
             int min_threshold = MIN_THRESHOLD;
             int min_length = StateMachine::MIN_LENGTH;
-            double marker_aspect_tolerance = MARKER_ASPECT_TOLERANCE;
-            double alpha = ALPHA;
-            double tolerance = StateMachine::TOLERANCE;
+            float marker_aspect_tolerance = MARKER_ASPECT_TOLERANCE;
+            float alpha = ALPHA;
+            float tolerance = StateMachine::TOLERANCE;
             std::vector<std::shared_ptr<Marker>> markers;
     };
 }
