+++ src/include/PersistentTarget.hpp
@@ -13,27 +13,27 @@ namespace targetfinder {
         friend class Target;
         friend class TargetFinder;
     public:
-        void update(Target *new_target, int64 timestamp, double influence);
+        void update(Target *new_target, int64 timestamp, float influence);
         bool alive(int64 timestamp);
-        double similarity(Target *other);
-        double angle();
+        float similarity(Target *other);
+        float angle();
         cv::Rect rect();
         cv::Point center();
         void setTimeout(int64 ticks);
-        void setAngleOffset(double angle);
+        void setAngleOffset(float angle);
         std::string str();
         int age();
-        double distance(Target *other);
-        double velocity();
+        float distance(Target *other);
+        float velocity();
         void tick();
     protected:
         int64 timeout_ticks;
         int64 last_updated;
         int lifetime;
         int x, y, width, height;
-        double calc_angle, angle_offset;
+        float calc_angle, angle_offset;
         bool calc_alive = true;
-        double calc_velocity = 0.0;
+        float calc_velocity = 0.0;
     };
 
 }
